# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Developnow Test-Api Swagger Documentation",
        "description": "Developnow Api Swagger documentation description",
        "contact": {
            "email": "veshraj.joshi1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Developnow Test API Server"
        }
    ],
    "paths": {
        "/api/files": {
            "post": {
                "tags": [
                    "File Uploads"
                ],
                "summary": "Store new Files",
                "description": "Returns project Files",
                "operationId": "api/v1/files-create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "File to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "status",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "Item Created successfully",
                                            "type": "string"
                                        },
                                        "payload": {
                                            "$ref": "#/components/schemas/Image"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/images": {
            "post": {
                "tags": [
                    "Image Upload"
                ],
                "summary": "Store new Images",
                "description": "Returns Image Url",
                "operationId": "api/v1/images-create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Image to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "status",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "Item Created successfully",
                                            "type": "string"
                                        },
                                        "payload": {
                                            "$ref": "#/components/schemas/Image"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Forbidden": {
                "title": "Forbidden",
                "description": "Forbidden model",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "format": "string",
                        "example": "err"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "format": "string",
                        "example": "Either this content doesnot exists or you do not have enough priviledges"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Forbidden"
                }
            },
...
            "NotFound": {
                "title": "NotFound",
                "description": "NotFound model",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "format": "string",
                        "example": "err"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "format": "string",
                        "example": "404 Page not found"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NotFound"
                }
            },
...
            "Unauthenticated": {
                "title": "Unauthenticated",
                "description": "Unauthenticated model",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "format": "string",
                        "example": "err"
                    },
                    "message": {
                        "title": "message",
                        "description": "message",
                        "type": "string",
                        "format": "string",
                        "example": "You are not logged in. Please login first."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Unauthenticated"
                }
            },
            "Image": {
                "title": "Images",
                "description": "Images model",
                "properties": {
                    "imageUrl": {
                        "title": "imageUrl",
                        "description": "imageUrl",
                        "type": "string",
                        "example": "http://xyz.com/abc.jpg"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Images"
                }
            },
...
        }
    }
}